git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@15347 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/prolifegql/CWorkspaceControllerComp.cpp
@@ -981,7 +981,7 @@ QString CWorkspaceControllerComp::GenerateColorFromString(const QString& text) c
 		return QString("#CCCCCC");
 	}
 
-	size_t hash = qHash(text);
+	uint hash = qHash(text);
 	int index = static_cast<int>(hash % s_standardColors.size());
 	return s_standardColors[index];
 }
